Add aliases for VTK targets.
For VTK-8, add VTK_INCLUDE_DIRS to all the targets for which an alias was created. Add a few more vtk libs that are used by VisIt directly to the required list.
Remove use of VTK_INCLUDE_DIRS.
Replace use of vtk targets with the aliased versions. Removed a lot of VTK_VERSION tests which aren't needed due to use of the alias.